+++ uclr_acquisition/sensors/__init__.py
@@ -1,5 +1,7 @@
 from .usg import USGScanner
+from .mems import MEMSMicrophone
 
 usg_scanner = None
+mems_microphone = None
 
-__all__ = ['USGScanner', 'usg_scanner']
+__all__ = ['USGScanner', 'MEMSMicrophone', 'usg_scanner', 'mems_microphone']